pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Urkel",email = "urkelcodes7@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "djoser (>=2.3.1,<3.0.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)",
    "django-cors-headers (>=4.6.0,<5.0.0)",
    "drf-yasg (>=1.21.8,<2.0.0)",
    "whitenoise (>=6.9.0,<7.0.0)",
    "dj-database-url (>=2.3.0,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
